requirements.txt
asgiref==3.6.0
Django==4.1.5
django-dotenv==1.4.2
numpy==1.24.1
pandas==1.5.2
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
